+++ src/models/scrapes.js
@@ -1,5 +1,5 @@
+// returns all data that match scrapeUrl
 function getScrape(client, url) {
-    // TODO: Rewrite to use promises vs callback
     return new Promise(function (resolve, reject) {
         client.query('SELECT * FROM crawls where raw_url = $1', [url], function (err, response) {
             if (err) {
@@ -11,13 +11,14 @@ function getScrape(client, url) {
     })
 }
 
-// TODO: Add addScrape
+// adds data to DB
 function addScrape(client, url, data) {
 
     client.query('INSERT INTO crawls (raw_url, title, description, image, favicon) VALUES ($1, $2, $3, $4, $5)', [url, data.title, data.description, data.image, data.favicon])
 
 }
 
+// updates row (rather than making a new row) in DB
 function updateScrape(client, url, data) {
     client.query('UPDATE crawls SET title = $1, description = $2, image = $3, favicon = $4 WHERE raw_url = $5', [data.title, data.description, data.image, data.favicon, url])
 }